refactor: use Clerk billing types
