+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/(manage)/settings/billing/_components/loading-line.tsx
@@ -1,0 +1,10 @@
+import { Loader2 } from "lucide-react";
+
+export function LoadingLine({ label }: { label: string }) {
+  return (
+    <div className="mt-4 flex items-center gap-2 text-muted-foreground text-sm">
+      <Loader2 className="size-4 animate-spin" />
+      {label}
+    </div>
+  );
+}
