+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/(manage)/settings/billing/page.tsx
@@ -1,14 +1,7 @@
-import { HydrateClient, prefetch, trpc } from "@repo/app-trpc/server";
 import { BillingSettingsClient } from "./_components/billing-settings-client";
 
 export const dynamic = "force-dynamic";
 
 export default function BillingPage() {
-  prefetch(trpc.pendingNotAllowed.orgBilling.overview.queryOptions());
-
-  return (
-    <HydrateClient>
-      <BillingSettingsClient />
-    </HydrateClient>
-  );
+  return <BillingSettingsClient />;
 }
