+++ packages/app-validation/src/index.ts
@@ -6,5 +6,4 @@ export * from "./forms/team-form";
 export * from "./primitives/slugs";
 export * from "./schemas/activities";
 export * from "./schemas/org-api-key";
-export * from "./schemas/org-billing";
 export * from "./schemas/org-member";
