+++ packages/app-validation/src/schemas/index.ts
@@ -7,5 +7,4 @@
 
 export * from "./activities";
 export * from "./org-api-key";
-export * from "./org-billing";
 export * from "./org-member";
