--- packages/app-validation/src/schemas/org-billing.ts
@@ -1,5 +1,0 @@
-import { z } from "zod";
-
-export const cancelOrgBillingSubscriptionItemSchema = z.object({
-  subscriptionItemId: z.string().min(1),
-});
